Create Stock Quote Retrieval
A website that allows a user to retrieve stock data on inputting the stock symbol of a company. This is done using the Stock Quote API. The results are displayed in tabular format and charts format using HighCharts. HTML, CSS, Javascript, Ajax and PHP were used in this project for retrieving the Stock data, parsing the JSON and XML data, and displaying it in the required format.